refactor: tokenprovider interface화 및 DIP적용
