+++ src/main/kotlin/io/hhplus/concertreservationservice/application/helper/TokenProvider.kt
@@ -1,0 +1,7 @@
+package io.hhplus.concertreservationservice.application.helper
+
+interface TokenProvider {
+    fun generateToken(userId: Long): String
+
+    fun validateToken(token: String): Boolean
+}
